Rename lib container to containers
