+++ src/ystdlib/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_subdirectory(container)
+add_subdirectory(containers)
 add_subdirectory(error_handling)
