+++ taskfile.yaml
@@ -17,7 +17,7 @@ vars:
   G_TEST_BIN_DIR: "{{.G_BUILD_DIR}}/testbin"
   G_TEST_TARGET_SUFFIXES:
     - "all"
-    - "container"
+    - "containers"
     - "error_handling"
 
 tasks:
